Create custom element item-list-entry-dropdown
